move spinner.py to elements/
