+++ lib/streamlit/runtime/caching/cache_utils.py
@@ -40,6 +40,7 @@ from google.protobuf.message import Message
 import streamlit as st
 from streamlit import util
 from streamlit.elements import NONWIDGET_ELEMENTS
+from streamlit.elements.spinner import spinner
 from streamlit.logger import get_logger
 from streamlit.proto.Block_pb2 import Block
 from streamlit.runtime.caching.cache_errors import (
@@ -253,7 +254,7 @@ def create_cache_wrapper(cached_func: CachedFunction) -> Callable[..., Any]:
             return return_value
 
         if cached_func.show_spinner or isinstance(cached_func.show_spinner, str):
-            with st.spinner(message):
+            with spinner(message):
                 return get_or_create_cached_value()
         else:
             return get_or_create_cached_value()
